+++ src/pages/PeoplePage.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
 import { Loader } from '../components/Loader';
 import { getPeople } from '../api';
-import { PersonLink } from '../components/PersonLink';
 import { Person } from '../types';
+import { PeopleTable } from '../components/PeopleTable';
 
 export const PeoplePage = () => {
   const [people, setPeople] = useState<Person[] | null>([]);
@@ -19,10 +19,6 @@ export const PeoplePage = () => {
       .finally(() => setLoadingPeople(false));
   }, []);
 
-  const findEqualName = (parentName: string | null) => {
-    return people?.find(person => person.name === parentName);
-  };
-
   return (
     <>
       <h1 className="title">People Page</h1>
@@ -36,32 +32,8 @@ export const PeoplePage = () => {
             </p>
           )}
 
-          {!loadingPeople && !errorMessage && (
-            <table
-              data-cy="peopleTable"
-              className="table is-striped is-hoverable is-narrow is-fullwidth"
-            >
-              <thead>
-                <tr>
-                  <th>Name</th>
-                  <th>Sex</th>
-                  <th>Born</th>
-                  <th>Died</th>
-                  <th>Mother</th>
-                  <th>Father</th>
-                </tr>
-              </thead>
-
-              <tbody>
-                {people?.map(person => (
-                  <PersonLink
-                    key={person.slug}
-                    person={person}
-                    findEqualName={findEqualName}
-                  />
-                ))}
-              </tbody>
-            </table>
+          {!loadingPeople && !errorMessage && people && (
+            <PeopleTable people={people} />
           )}
 
           {!loadingPeople && !errorMessage && people?.length === 0 && (
